gap: Move platform creation to a seperate module
