+++ libs/gap/CMakeLists.txt
@@ -7,6 +7,7 @@ message(STATUS "> library: gap")
 add_library(volo_gap STATIC
   src/input.c
   src/pal.c
+  src/platform.c
   src/register.c
   src/window.c)
 target_include_directories(volo_gap PUBLIC include)
