+++ libs/gap/src/platform.c
@@ -1,0 +1,44 @@
+#include "ecs_world.h"
+
+#include "platform_internal.h"
+
+ecs_comp_define_public(GapPlatformComp);
+
+static void ecs_destruct_platform_comp(void* data) {
+  GapPlatformComp* comp = data;
+  gap_pal_destroy(comp->pal);
+}
+
+ecs_view_define(GapPlatformView) { ecs_access_write(GapPlatformComp); };
+
+ecs_system_define(GapPlatformUpdateSys) {
+
+  // Retrieve or create the global platform component.
+  EcsIterator*     platformItr = ecs_view_itr_first(ecs_world_view_t(world, GapPlatformView));
+  GapPlatformComp* platform;
+  if (platformItr) {
+    platform = ecs_view_write_t(platformItr, GapPlatformComp);
+  } else {
+    platform = ecs_world_add_t(
+        world,
+        ecs_world_entity_create(world),
+        GapPlatformComp,
+        .pal = gap_pal_create(g_alloc_heap));
+  }
+
+  // Process all os events.
+  gap_pal_update(platform->pal);
+}
+
+ecs_module_init(gap_platform_module) {
+  ecs_register_comp(GapPlatformComp, .destructor = ecs_destruct_platform_comp);
+
+  ecs_register_view(GapPlatformView);
+
+  /**
+   * NOTE: Create the system with thread-affinity as some platforms use thread-local event-queues so
+   * we need to serve them from the same thread every time.
+   */
+  ecs_register_system_with_flags(
+      GapPlatformUpdateSys, EcsSystemFlags_ThreadAffinity, ecs_view_id(GapPlatformView));
+}
