+++ libs/gap/src/platform_internal.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "ecs_module.h"
+
+#include "pal_internal.h"
+
+ecs_comp_extern_public(GapPlatformComp) { GapPal* pal; };
