+++ libs/gap/src/register.c
@@ -1,4 +1,7 @@
 #include "ecs_def.h"
 #include "gap_register.h"
 
-void gap_register(EcsDef* def) { ecs_register_module(def, gap_window_module); }
+void gap_register(EcsDef* def) {
+  ecs_register_module(def, gap_platform_module);
+  ecs_register_module(def, gap_window_module);
+}
